Fix signing
Sign and verify were not working because Sign was creating a 65 byte
signature with a 0 or 1 recovery id and Verify expected a 64 byte
signature with no recovery id.

GenerateProofOfPossessionSignature was also not functioning correctly
because the sign logic was failing to modify the signature recovery byte
to be 27 or 28 which is expected by the ecrecover evm contract.

Now Sign and Verify have been updated to produce and consume 64 byte
ecdsa signatures without a recovery byte.

GenerateProofOfPossessionSignature now correctly sets the recovery byte
to be either 27 or 28.

Also removed VerifyWithPrefix as it was not used.