Update IpcFile path
The docs for geth say that the ipc file if provided as a plain file name
is placed in the datadir, this was not happening. This change ensures
that if a plain file is specified for the geth ipc then it is stored in
the datadir. Although the "datadir" for some reason has '/celo' appended
to it by the rosetta server, so actually the ipc will appear there.