Update readme with instructions on releasing
Also modified the release script to not create and push a release
commit.

Since the master branch is protected in this repo, it's not possible to
push a commit directly to master, that means that any commit made by the
script would be on a branch, and given that we may squash and merge
those branches we could end up not having the tags on master, even
though the state of master could be identical to the tagged commit on
the branch.

The instructions now ensure that we first merge all release changes to
master, and then tag that.